+++ MoravaEngine/SceneAnimPBR.cpp
@@ -172,10 +172,7 @@ void SceneAnimPBR::ResizeViewport(glm::vec2 viewportPanelSize)
         m_RenderFramebuffer->Resize((uint32_t)viewportPanelSize.x, (uint32_t)viewportPanelSize.y);
         m_ViewportSize = glm::vec2(viewportPanelSize.x, viewportPanelSize.y);
 
-        m_CameraController->OnResize(viewportPanelSize.x, viewportPanelSize.y);
-
-        //  Log::GetLogger()->info("SceneAnimPBR::ResizeViewport Viewport Width: {0} Viewport Height: {1}, AspectRatio: {2}",
-        //      viewportPanelSize.x, viewportPanelSize.y, m_CameraController->GetAspectRatio());
+        m_CameraController->OnResize((uint32_t)viewportPanelSize.x, (uint32_t)viewportPanelSize.y);
     }
 }
